Collapse all rules + tests into dedicated plug-ins.
