--- Src/Plugins/SEC1001.SecurePlaintextSecrets/SEC1001.SecurePlaintextSecrets.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(SolutionDir)..\Targets\build.lib.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.product.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.plugins.props" />
-
-  <ItemGroup>
-    <Content Include="SEC1001.SecurePlaintextSecrets.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="build\SEC1001.SecurePlaintextSecrets.targets">
-      <PackagePath>build\</PackagePath>
-    </Content>
-  </ItemGroup>
-  
-</Project>
