--- Src/Plugins/SEC1002.ReviewPotentiallySensitiveData/SEC1002.ReviewPotentiallySensitiveData.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(SolutionDir)..\Targets\build.lib.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.product.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.plugins.props" />
-  <ItemGroup>
-    <Content Include="build\SEC1002.ReviewPotentiallySensitiveData.targets">
-      <PackagePath>build\</PackagePath>
-    </Content>
-  </ItemGroup>
-  <ItemGroup>
-    <None Update="SEC1002.ReviewPotentiallySensitiveData.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-  
-</Project>
