--- Src/Plugins/SEC1002.ReviewPotentiallySensitiveData/SEC1002.ReviewPotentiallySensitiveData.json
@@ -1,25 +1,0 @@
-{
-  "Definitions": [
-    {
-      "Name": "ReviewPotentiallySensitiveData",
-      "Id": "SEC1002",
-      "Level": "Note",
-      "FileNameAllowRegex": "(?i)\\.(azure|bat|c|cmd|config|cpp|cs|cscfg|definitions|dtsx|h|hxx|hpp|ini|java|jsx?|json|keys|kt|loadtest|m|md|php|properties|ps1|psm1|pubxml|py|resx|sample|sql|ste|swift|test|tsx?|txt|waz|xml)$",
-      "Description": "Review exposed potentially sensitive data, such as service principal names, social security numbers, etc.",
-      "Message": "'{0:scanTarget}' contains an apparent {1:dataKind}.",
-      "MatchExpressions": [
-        {
-          "SubId": "EmailAddress",
-          "ContentsRegex": "(?i)0-9a-z._%+-]+@[0-9a-z.-]+\\.[a-z]{2,4}",
-          "MessageArguments": { "dataKind": "email address" },
-          "Notes": [ "We should look for a more complex/precise regex for this data kind." ]
-        },
-        {
-          "SubId": "SocialSecurityNumber",
-          "ContentsRegex": "\\b[0-9]{3}-[0-9]{2}-[0-9]{4}\\b",
-          "MessageArguments": { "dataKind": "social security number" }
-        }
-      ]
-    }
-  ]
-}
