--- Src/Plugins/SEC1003.ReviewPotentiallySensitiveFiles/SEC1003.ReviewPotentiallySensitiveFiles.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(SolutionDir)..\Targets\build.lib.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.product.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.plugins.props" />
-
-  <ItemGroup>
-    <Content Include="SEC1003.ReviewPotentiallySensitiveFiles.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="build\SEC1003.ReviewPotentiallySensitiveFiles.targets">
-      <PackagePath>build\</PackagePath>
-    </Content>
-  </ItemGroup>
-
-</Project>
