--- Src/Plugins/SEC1003.ReviewPotentiallySensitiveFiles/SEC1003.ReviewPotentiallySensitiveFilesValidators.cs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-
-namespace Microsoft.CodeAnalysis.SarifPatternMatcher.SensitiveFiles
-{
-    public class SensitiveFilesValidators
-    {
-    }
-}
