--- Src/Plugins/SEC1004.UseSecureApi/SEC1004.UseSecureApi.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(SolutionDir)..\Targets\build.lib.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.product.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.plugins.props" />
-
-  <ItemGroup>
-    <Content Include="SEC1004.UseSecureApi.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="build\SEC1004.UseSecureApi.targets">
-      <PackagePath>build\</PackagePath>
-    </Content>
-  </ItemGroup>
-
-</Project>
