--- Src/Plugins/SEC1004.UseSecureApi/SEC1004.UseSecureApiValidators.cs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-
-namespace Microsoft.CodeAnalysis.SarifPatternMatcher.BannedApi
-{
-    public class BannedApiValidators
-    {
-    }
-}
