+++ Src/Plugins/Security/Security.csproj
@@ -5,7 +5,7 @@
   <Import Project="$(SolutionDir)..\Targets\build.plugins.props" />
 
   <ItemGroup>
-    <Content Include="SEC1001.SecurePlaintextSecrets.json">
+    <Content Include="*.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="build\SEC1001.SecurePlaintextSecrets.targets">
