--- Src/Plugins/Tests.PlaintextSecrets/TestData/PlaintextSecrets/ExpectedOutputs/VariousPats.sarif
@@ -1,112 +1,0 @@
-{
-  "$schema": "https://schemastore.azurewebsites.net/schemas/json/sarif-2.1.0-rtm.5.json",
-  "version": "2.1.0",
-  "runs": [
-    {
-      "tool": {
-        "driver": {
-          "name": "testhost",
-          "organization": "Microsoft Corporation",
-          "product": "Microsoft.TestHost",
-          "fullName": "testhost 15.0.0.0",
-          "version": "15.0.0.0",
-          "semanticVersion": "15.0.0",
-          "rules": [
-            {
-              "id": "SEC1001",
-              "name": "SearchSkimmer",
-              "fullDescription": {
-                "text": "Do not expose plaintext (or base64-encoded plaintext) secrets in versioned engineering content."
-              },
-              "messageStrings": {
-                "Default": {
-                  "text": "'{0}' contains an apparent {1} {2}."
-                }
-              },
-              "defaultConfiguration": {
-                "level": "error"
-              },
-              "helpUri": "https://github.com/microsoft/sarif-pattern-matcher"
-            }
-          ]
-        }
-      },
-      "invocations": [
-        {
-          "executionSuccessful": true
-        }
-      ],
-      "results": [
-        {
-          "ruleId": "SEC1001/AzureDevOpsPersonalAccessToken",
-          "ruleIndex": 0,
-          "level": "error",
-          "message": {
-            "id": "Default",
-            "arguments": [
-              "VariousPats.txt",
-              "base64-encoded",
-              "Azure DevOps personal access token"
-            ]
-          },
-          "locations": [
-            {
-              "physicalLocation": {
-                "artifactLocation": {
-                  "uri": "src/Plugins/Tests.PlaintextSecrets/TestData/PlaintextSecrets/Inputs/VariousPats.txt",
-                  "uriBaseId": "SRC_ROOT"
-                },
-                "region": {
-                  "startLine": 12,
-                  "startColumn": 1,
-                  "endLine": 12,
-                  "endColumn": 73,
-                  "charOffset": 442,
-                  "charLength": 72,
-                  "snippet": {
-                    "text": "aDVseGVxa3o0enVqa3l2emc1ZW1rZWpnZWx4ajR4NjR0dnl6Y2xwcmVrdWxvYXdiYno1cQ=="
-                  }
-                }
-              }
-            }
-          ]
-        },
-        {
-          "ruleId": "SEC1001/AzureDevOpsPersonalAccessToken",
-          "ruleIndex": 0,
-          "level": "error",
-          "message": {
-            "id": "Default",
-            "arguments": [
-              "VariousPats.txt",
-              "plaintext",
-              "Azure DevOps personal access token"
-            ]
-          },
-          "locations": [
-            {
-              "physicalLocation": {
-                "artifactLocation": {
-                  "uri": "src/Plugins/Tests.PlaintextSecrets/TestData/PlaintextSecrets/Inputs/VariousPats.txt",
-                  "uriBaseId": "SRC_ROOT"
-                },
-                "region": {
-                  "startLine": 9,
-                  "startColumn": 1,
-                  "endLine": 9,
-                  "endColumn": 53,
-                  "charOffset": 326,
-                  "charLength": 52,
-                  "snippet": {
-                    "text": "h5lxeqkz4zujkyvzg5emkejgelxj4x64tvyzclprekuloawbbz5q"
-                  }
-                }
-              }
-            }
-          ]
-        }
-      ],
-      "columnKind": "utf16CodeUnits"
-    }
-  ]
-}
