--- Src/Plugins/Tests.PotentiallySensitiveData/TestData/EmptyResults.sarif
@@ -1,15 +1,0 @@
-{
-  "$schema": "https://schemastore.azurewebsites.net/schemas/json/sarif-2.1.0-rtm.4.json",
-  "version": "2.1.0",
-  "runs": [
-    {
-      "tool": {
-        "driver": {
-          "name": "Shell file for initializing tests."
-        }
-      },
-      "results": [],
-      "columnKind": "utf16CodeUnits"
-    }
-  ]
-}
