--- Src/Plugins/Tests.PotentiallySensitiveData/Tests.PotentiallySensitiveData.csproj
@@ -1,31 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(SolutionDir)..\Targets\build.app.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.test.props" />
-
-  <ItemGroup>
-    <EmbeddedResource Include="TestData\PotentiallySensitiveData\ExpectedOutputs\EmailAddresses.sarif">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </EmbeddedResource>
-    <EmbeddedResource Include="TestData\PotentiallySensitiveData\ExpectedOutputs\SocialSecurityNumbers.sarif">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </EmbeddedResource>
-    <EmbeddedResource Include="TestData\PotentiallySensitiveData\Inputs\EmailAddresses.txt">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </EmbeddedResource>
-    <EmbeddedResource Include="TestData\PotentiallySensitiveData\Inputs\SocialSecurityNumbers.txt">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </EmbeddedResource>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\sarif-sdk\src\Sarif.Driver\Sarif.Driver.csproj" />
-    <ProjectReference Include="..\..\sarif-sdk\src\Sarif\Sarif.csproj" />
-    <ProjectReference Include="..\..\sarif-sdk\src\Test.Utilities.Sarif\Test.Utilities.Sarif.csproj" />
-    <ProjectReference Include="..\..\SarifPatternMatcher\SarifPatternMatcher.csproj" />
-    <ProjectReference Include="..\PotentiallySensitiveData\PotentiallySensitiveData.csproj" />
-
-    <Compile Include="$(SolutionDir)..\src\Plugins\EndToEndTests.cs" />
-  </ItemGroup>
-
-</Project>
