+++ Src/Plugins/Tests.Security/TestData/EmptyResults.sarif
