+++ Src/Plugins/Tests.Security/TestData/ReviewPotentiallySensitiveData/Inputs/EmailAddresses.txt
