+++ Src/Plugins/Tests.Security/TestData/ReviewPotentiallySensitiveData/UnitTestOutput.PlaintextSecrets/ActualOutputs/VariousPats.sarif
