+++ Src/Plugins/Tests.Security/TestData/ReviewPotentiallySensitiveData/UnitTestOutput.PlaintextSecrets/ExpectedOutputs/VariousPats.sarif
