+++ Src/Plugins/Tests.Security/TestData/ReviewPotentiallySensitiveFiles/Inputs/Fake.asc
@@ -1,0 +1,10 @@
+ Volume in drive E is DATADRIVE1
+ Volume Serial Number is 1235-2302
+
+ Directory of E:\src\spam\Src\Plugins\Tests.Security\TestData\ReviewPotentiallySensitiveFiles\Inputs
+
+12/17/2020  11:40 AM    <DIR>          .
+12/17/2020  11:40 AM    <DIR>          ..
+12/17/2020  11:40 AM                 0 Fake.asc
+               1 File(s)              0 bytes
+               2 Dir(s)  11,484,848,128 bytes free
