+++ Src/Plugins/Tests.Security/TestData/SecurePlaintextSecrets/Inputs/VariousPats.txt
