+++ Src/Plugins/Tests.Security/TestData/UseSecureApi/ExpectedOutputs/Malloc.sarif
@@ -1,0 +1,104 @@
+{
+  "$schema": "https://schemastore.azurewebsites.net/schemas/json/sarif-2.1.0-rtm.5.json",
+  "version": "2.1.0",
+  "runs": [
+    {
+      "tool": {
+        "driver": {
+          "name": "testhost",
+          "organization": "Microsoft Corporation",
+          "product": "Microsoft.TestHost",
+          "fullName": "testhost 15.0.0.0",
+          "version": "15.0.0.0",
+          "semanticVersion": "15.0.0",
+          "rules": [
+            {
+              "id": "SEC1004/Memory/Allocation",
+              "name": "SearchSkimmer",
+              "fullDescription": {
+                "text": "Developers should use secure API in preferance of insecure alternates."
+              },
+              "messageStrings": {
+                "Default": {
+                  "text": "'{0}' contains a call to '{1}', a potentially insecure API that could be replaced with a more secure alternative: '{2}'."
+                }
+              },
+              "defaultConfiguration": {
+                "level": "error"
+              },
+              "helpUri": "https://github.com/microsoft/sarif-pattern-matcher"
+            }
+          ]
+        }
+      },
+      "invocations": [
+        {
+          "executionSuccessful": true
+        }
+      ],
+      "results": [
+        {
+          "ruleId": "SEC1004/Memory/Allocation/_alloca",
+          "ruleIndex": 0,
+          "level": "error",
+          "message": {
+            "id": "Default",
+            "arguments": [
+              "Malloc.c",
+              "_alloca",
+              "_malloca"
+            ]
+          },
+          "locations": [
+            {
+              "physicalLocation": {
+                "artifactLocation": {
+                  "uri": "src/Plugins/Tests.Security/TestData/UseSecureApi/Inputs/Malloc.c",
+                  "uriBaseId": "SRC_ROOT"
+                },
+                "region": {
+                  "startLine": 10,
+                  "startColumn": 14,
+                  "endLine": 10,
+                  "endColumn": 22,
+                  "charOffset": 182,
+                  "charLength": 8,
+                  "snippet": {
+                    "text": "_alloca("
+                  }
+                }
+              }
+            }
+          ],
+          "fixes": [
+            {
+              "description": {
+                "text": "Replace '_alloca with '_malloca'."
+              },
+              "artifactChanges": [
+                {
+                  "artifactLocation": {
+                    "uri": "src/Plugins/Tests.Security/TestData/UseSecureApi/Inputs/Malloc.c",
+                    "uriBaseId": "SRC_ROOT"
+                  },
+                  "replacements": [
+                    {
+                      "deletedRegion": {
+                        "charOffset": 182,
+                        "charLength": 8
+                      },
+                      "insertedContent": {
+                        "text": "_malloca("
+                      }
+                    }
+                  ]
+                }
+              ]
+            }
+          ]
+        }
+      ],
+      "columnKind": "utf16CodeUnits"
+    }
+  ]
+}
