--- Src/Plugins/Tests.SensitiveFiles/Tests.SensitiveFiles.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(SolutionDir)..\Targets\build.app.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.test.props" />
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\sarif-sdk\src\Sarif.Driver\Sarif.Driver.csproj" />
-    <ProjectReference Include="..\..\sarif-sdk\src\Sarif\Sarif.csproj" />
-    <ProjectReference Include="..\..\sarif-sdk\src\Test.Utilities.Sarif\Test.Utilities.Sarif.csproj" />
-    <ProjectReference Include="..\..\SarifPatternMatcher\SarifPatternMatcher.csproj" />
-    <ProjectReference Include="..\SensitiveFiles\SensitiveFiles.csproj" />
-
-    <Compile Include="$(SolutionDir)..\src\Plugins\EndToEndTests.cs" />
-  </ItemGroup>
-  
-</Project>
